requirements.txt
altair==4.2.0
argh==0.26.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
astor==0.8.1
asttokens==2.0.5
attrs==21.4.0
backcall==0.2.0
backports.functools-lru-cache==1.6.4
base58==2.1.1
black==22.1.0
bleach==4.1.0
blinker==1.4
boto3==1.20.54
botocore==1.23.54
Brotli==1.0.9
brotlipy==0.7.0
cachetools==5.0.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.11
click==8.0.3
colorama==0.4.4
cryptography==36.0.1
cycler==0.11.0
dash==2.1.0
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-renderer==1.9.1
dash-table==5.0.0
dataclasses==0.6
debugpy==1.5.1
decorator==5.1.1
defusedxml==0.7.1
entrypoints==0.4
executing==0.8.2
Flask==2.0.2
Flask-Compress==1.10.1
flit_core==3.6.0
fonttools==4.29.1
gitdb==4.0.9
GitPython==3.1.26
greenlet==1.1.2
idna==3.3
importlib-metadata==4.11.0
importlib-resources==5.4.0
ipykernel==6.9.0
ipython==7.31.1
ipython-genutils==0.2.0
ipywidgets==7.6.5
itsdangerous==2.0.1
jedi==0.18.1
Jinja2==3.0.3
jmespath==0.10.0
joblib==1.1.0
jsonschema==4.4.0
jupyter-client==7.1.2
jupyter-core==4.9.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.2
kiwisolver==1.3.2
MarkupSafe==2.0.1
matplotlib==3.5.1
matplotlib-inline==0.1.3
mistune==0.8.4
munkres==1.1.4
mypy-extensions==0.4.3
nbclient==0.5.10
nbconvert==6.4.2
nbformat==5.1.3
nest-asyncio==1.5.4
notebook==6.4.8
numpy==1.21.5
packaging==21.3
pandas==1.3.5
pandocfilters==1.5.0
parso==0.8.3
pathspec==0.9.0
pickleshare==0.7.5
Pillow==9.0.1
pip==22.0.3
platformdirs==2.5.0
plotly==5.6.0
prometheus-client==0.13.1
prompt-toolkit==3.0.27
protobuf==3.18.1
pure-eval==0.2.2
pyarrow==6.0.0
#pycairo==1.20.1
pycparser==2.21
pydeck==0.7.1
Pygments==2.11.2
Pympler==0.9
pyOpenSSL==22.0.0
pyparsing==3.0.7
pyrsistent==0.18.1
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
#pywin32==303
#pywinpty==2.0.2
PyYAML==6.0
pyzmq==22.3.0
reportlab==3.5.68
requests==2.27.1
s3transfer==0.5.1
scikit-learn==1.0.2
scipy==1.7.3
Send2Trash==1.8.0
setuptools==60.8.2
six==1.16.0
smmap==3.0.5
SQLAlchemy==1.4.31
stack-data==0.1.4
streamlit==1.5.1
streamlit-plotly-events==0.0.6
tenacity==8.0.1
terminado==0.13.1
testpath==0.5.0
threadpoolctl==3.1.0
toml==0.10.2
tomli==2.0.1
toolz==0.11.2
torch==1.10.2
tornado==6.1
traitlets==5.1.1
typed-ast==1.5.2
typing_extensions==4.0.1
tzdata==2021.5
tzlocal==4.1
unicodedata2==14.0.0
urllib3==1.26.8
validators==0.18.2
watchdog==2.1.6
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.0.3
wheel==0.37.1
widgetsnbextension==3.5.2
win-inet-pton==1.1.0
zipp==3.7.0
rdkit-pypi
dgl
dgllife
wget==3.2
